Add wifi_syslog readme

+++ porters/wifi_syslog.sh
@@ -9,6 +9,11 @@ function die () {
     exit 1
 }
 
+# Check permission
+if [ `whoami` != 'omnilab' ]; then
+    die "Need permission of OMNILAB to run. Try user omnilab."
+fi
+
 # Global vars
 BASEDIR=$(dirname $0)/..
 source $BASEDIR/global_config.sh
@@ -17,6 +22,17 @@ source $BASEDIR/global_config.sh
 BINJAR=$BASEDIR/etlers/ArubaSyslog/target/scala-2.10/ArubaSyslog-assembly-1.0.jar
 CLSNAME="cn.edu.sjtu.omnilab.odh.spark.CleanseWifiLogs"
 
+# Check root path for raw data
+if [ ! -d $WIFI_SYSLOG_PATH ]; then
+    die "Cann't find path for raw syslog: $WIFI_SYSLOG_PATH"
+fi
+
+# Check root path for this repo
+if ! hadoop fs -test -d $HDFS_WIFI_SYSLOG; then
+    hadoop fs -mkdir -p $HDFS_WIFI_SYSLOG
+fi
+
+# Clear temporary folder
 TEMPWP=$HDFS_WIFI_SYSLOG/_temp
 hadoop fs -rm -r $TEMPWP
 hadoop fs -mkdir -p $TEMPWP
@@ -32,18 +48,19 @@ for rawfile in `ls $WIFI_SYSLOG_PATH`; do
 
     if [ $year$month$day == $(date -d "yesterday" '+%Y%m%d') ]; then
 
-        if ! hadoop fs -test -e $HDFS_WIFI_SYSLOG/$rfname/_SUCCESS; then
+	if ! hadoop fs -test -e $HDFS_WIFI_SYSLOG/$rfname/_SUCCESS; then
 
-            # Decompress file
-            if ! hadoop fs -test -e $TEMPWP/$rfname; then
-                gunzip -c $WIFI_SYSLOG_PATH/$rawfile | hadoop fs -put - $TEMPWP/$rfname
-            fi
+	    # Decompress file
+	    if ! hadoop fs -test -e $TEMPWP/$rfname; then
+		gunzip -c $WIFI_SYSLOG_PATH/$rawfile | hadoop fs -put - $TEMPWP/$rfname
+	    fi
 
-            # Cleanse wifilog
-            hadoop fs -rm -r $HDFS_WIFI_SYSLOG/$rfname
-            spark-submit2 --class $CLSNAME $BINJAR $TEMPWP/$rfname $HDFS_WIFI_SYSLOG/$rfname
+	    # Cleanse wifilog
+	    hadoop fs -rm -r $HDFS_WIFI_SYSLOG/$rfname
+	    spark-submit2 --class $CLSNAME $BINJAR $TEMPWP/$rfname $HDFS_WIFI_SYSLOG/$rfname
+	    hadoop fs -rm -r $TEMPWP/$rfname
+	fi
 
-        fi
     fi
 
 done
